fix tag set
* read hi lo and pal into new tag

* decompose tag in draw

* format as 8 16 8 - lo hi pal

* fix pal read for sectors

* comments on wall tiles

* refactor tags to pack via union

* rollback temp changes

* fix display of values

* fix setting of tags

+++ build2vs/build2.c
@@ -8187,7 +8187,7 @@ static void executepack (unsigned char *recvbuf, int doplaysound)
 								if ((unsigned)gps->grabsect < (unsigned)gst->numsects)
 								{
 									char *ptr = &cptr[5];
-									long lotag = 0, hitag = 0, pal = 0;
+									short lotag = 0, hitag = 0, pal = 0;
 
 									// Parse lotag
 									lotag = atol(ptr);
